console: Refactor device importer
